Version 1.31
Incorporated Guillaume Lafage's changes for link resolution and
improvements for the 'list checkouts' and 'list activities' windows.
Now double-click or Enter on a name in the list checkouts window will split-open that file, and double-click or Enter on an activity will set the current activity to be that of the activity under the cursor.
Some updates to the help file, added a thanks section to recognize those who have contributed to ccase.vim (thanks
everyone, I'm still open to more suggestions).