Version 1.36
My changes:
Add User commands for all the regular cabbrevs used, so that diff commands and others may be issued on the command line with '-c'.

Changes from Gary Johnson and WEN Guopeng:
Addition of a diff with first version on this branch.

Stefan Berglund and WEN Guopeng:
Alter 'ctpwv' to not use the $view variable, and instead get it from the system.

Changes from WEN Guopeng:
Changes to embed the ccase.vim documentation in the .vim file, and automatically install on startup, diff with merged version, and other wonderful addititions that I'm probably forgetting.